chore: refactor upvote
