+++ packages/frontend/src/components/example/Upvote.tsx
@@ -1,35 +1,6 @@
 import { useCallback, useState } from 'react';
-import axiosBase from 'axios';
-import { DOMAIN, UPVOTE_TYPES } from '@personaelabs/nymjs';
 import { useSignTypedData } from 'wagmi';
-import { getLatestGroup } from '@/lib/example-utils';
-
-const axios = axiosBase.create({
-  baseURL: `/api/v1`,
-});
-
-export const submitUpvote = async (postId: string, signTypedDataAsync: any) => {
-  const group = await getLatestGroup();
-
-  const upvote = {
-    postId: postId,
-    groupRoot: group.root,
-    timestamp: Math.round(Date.now() / 1000),
-  };
-
-  const sig = await signTypedDataAsync({
-    primaryType: 'Upvote',
-    domain: DOMAIN,
-    types: UPVOTE_TYPES,
-    message: upvote,
-  });
-
-  await axios.post(`/posts/${postId}/upvote`, {
-    groupRoot: upvote.groupRoot,
-    timestamp: upvote.timestamp,
-    sig,
-  });
-};
+import { submitUpvote } from '@/lib/actions';
 
 const UpvoteExample = () => {
   const [postId, setPostId] = useState<string>('');
